fix: use int32 cast in filterRGBV3 to handle uint32 underflow
This is a 32-bit vs 64-bit portability bug. The author's guard:

    if posR < 0 || width < 0 { return buf, nil }

proves they intended to catch the case where r[0] < 3 makes width
negative. On 32-bit Go (int is 32 bits), int(uint32(0)-3) = -3
(negative, caught by the guard). On 64-bit Go (int is 64 bits),
int(uint32(0)-3) = int(0xFFFFFFFD) = 4294967293 (positive, bypasses
the guard entirely, causing massive out-of-bounds access).

The fix uses int32 as an intermediate cast so the uint32 underflow
wraps to a negative value before widening to int, making the author's
original guard work correctly on both 32-bit and 64-bit platforms.

+++ filters.go
@@ -153,7 +153,7 @@ func abs(n int) int {
 }
 
 func filterRGBV3(r map[int]uint32, global, buf []byte, offset int64) ([]byte, error) {
-	width := int(r[0] - 3)
+	width := int(int32(r[0]) - 3)
 	posR := int(r[1])
 	if posR < 0 || width < 0 {
 		return buf, nil
